pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.2",
    "google-genai>=1.27.0",
    "pydantic>=2.11.7",
    "sift-stack-py>=0.7.0",
]
